+++ packages/logger/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @typed-assistant/logger
+
+## 0.0.22
+
+### Patch Changes
+
+- Limit log.txt size to 3MB.