+++ packages/utils/CHANGELOG.md
@@ -1,2 +1,8 @@
 # @typed-assistant/utils
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @typed-assistant/logger@0.0.22